Move notation test classes to correct package
Move the notation test classes to the package that
corresponds to the actual classes location.
