+++ src/test/java/org/wmn4j/io/musicxml/MusicXmlReaderDomTest.java
@@ -5,7 +5,7 @@ package org.wmn4j.io.musicxml;
 
 import org.junit.jupiter.api.Test;
 import org.wmn4j.io.ParsingFailureException;
-import org.wmn4j.notation.TestHelper;
+import org.wmn4j.TestHelper;
 import org.wmn4j.notation.ScoreBuilder;
 import org.wmn4j.notation.Score;
