+++ src/test/java/org/wmn4j/mir/discovery/GeometricPatternDiscoveryTest.java
@@ -6,7 +6,7 @@ package org.wmn4j.mir.discovery;
 import org.junit.jupiter.api.Test;
 import org.wmn4j.mir.Pattern;
 import org.wmn4j.mir.PatternPosition;
-import org.wmn4j.notation.TestHelper;
+import org.wmn4j.TestHelper;
 import org.wmn4j.notation.Durations;
 import org.wmn4j.notation.Note;
 import org.wmn4j.notation.Pitch;
