+++ src/test/java/org/wmn4j/mir/discovery/PointSetTest.java
@@ -1,7 +1,7 @@
 package org.wmn4j.mir.discovery;
 
 import org.junit.jupiter.api.Test;
-import org.wmn4j.notation.TestHelper;
+import org.wmn4j.TestHelper;
 import org.wmn4j.notation.Durations;
 import org.wmn4j.notation.Pitch;
 import org.wmn4j.notation.Score;
