+++ src/test/java/org/wmn4j/notation/RestTest.java
@@ -1,7 +1,7 @@
 /*
  * Distributed under the MIT license (see LICENSE.txt or https://opensource.org/licenses/MIT).
  */
-package org.wmn4j.notation.elements;
+package org.wmn4j.notation;
 
 import org.junit.jupiter.api.Test;
 import org.wmn4j.notation.Durations;
